added terrain

+++ src/model/Player.java
@@ -1,7 +1,3 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
 package model;
 
 /**
@@ -9,33 +5,30 @@ package model;
  * @author Joebh
  */
 public class Player {
-    
-    private double x, y;
+
+    private float x, y;
 
     public Player() {
     }
-    
-    public Player(double x, double y) {
+
+    public Player(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
-    
-    
-    
 }
